Move InvitationsPost and MyInvitationsGet to invitations.py.
